Move encrypt/decrypt into crypto/encryption.ts
